add: courses router

+++ src/models/user/user.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import TEXT, TIMESTAMP, BigInteger, String
+from sqlalchemy import TEXT, TIMESTAMP, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src.models.base import Base
@@ -9,7 +9,7 @@ from src.models.base import Base
 class User(Base):
     __tablename__ = "User"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
     email: Mapped[str] = mapped_column(String(255), nullable=False, unique=True)
     hashed_password: Mapped[str] = mapped_column(TEXT, nullable=False)
